% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_seurat.R
\name{make_seurat}
\alias{make_seurat}
\title{Make Seurat object}
\usage{
make_seurat(cellranger_path, project_id, hto_names = NULL, mito_ids, ribo_ids)
}
\arguments{
\item{cellranger_path}{Path to the CellRanger filtered feature matrix directory.}

\item{project_id}{A name for your project.}

\item{hto_names}{A vector for naming each hashtag, in the same order as in the feature matrix.}

\item{mito_ids}{A vector of ENSEMBL gene IDs for mitochondrial genes.}

\item{ribo_ids}{A vector of ENSEMBL gene IDs for ribosomal protein genes.}
}
\value{
A Seurat Object
}
\description{
Make Seurat object
}
\examples{
my_obj <- make_seurat("filtered_feature_bc_matrix", "PBMCs", c("condition1", "condition2"), mito_genes_ids, ribo_genes_ids)
}
