% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_admix.R
\name{plot_admix}
\alias{plot_admix}
\title{Plot Admixture-style plots of HTO proportions in droplets}
\usage{
plot_admix(seurat_obj, colors_vec)
}
\arguments{
\item{seurat_obj}{A Seurat object after HTODemux was run.}

\item{color_vec}{A named vector of custom colors, with HTOs as names.}
}
\value{
A ggplot.
}
\description{
Plot Admixture-style plots of HTO proportions in droplets
}
\examples{
mycolors <- rainbow(4)
names(mycolors) <- c("black", "green", "blue", "tomato3")
plot_admix(pbmc.3k, colors_vec = rainbow(8))
}
