% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_markers.R
\name{plot_markers}
\alias{plot_markers}
\title{Plot marker genes}
\usage{
plot_markers(seurat_obj, cluster_df)
}
\arguments{
\item{seurat_obj}{A Seurat object.}

\item{cluster_df}{A data.frame such as returned by Seurat::FindAllMarkers with added gene_id and gene_name columns.}
}
\value{
A dotplot of the top 10 marker genes per identity.
}
\description{
Plot marker genes
}
